[IOTMID-185] Initialization updates
This commit adds the current version of the calls required to make
kong work with the set of internal services.

This also adds a new set of configuration to the yaml file, easing
the issue where services (containers) would error and fail (exit)
due to non-initialized dependencies. This is actually just a hot
fix, since those services (notably, STH and Kong) should be able
to recover from dependency errors on their own. On a more elaborate
deploy scenario (i.e. using kubernetes or mesos instead of docker-compose)
the "hypervisor" can be configured to reload failed containers on
error.

Change-Id: I49a4498a9b63dd6840e0ba92528a58dcdabb8d30