[IOTMID-884] Adds missing configuration for HTTP binding on iotagent-json
This adds the missing http section from the agent's config, allowing
devices to use the HTTP interface of iotagent to publish attribute
data.

Also, this adds a new endpoint configuration, to allow kong to proxy
the attribute updates to the agent.

That said, to update device data using HTTP:

Issue a POST to:

   http://<middleware address>/iot/d?i=<device id>&k=<user's service>

With a JSON payload identical to the one used for MQTT.

Change-Id: I3de5de45a5a31bacab8248ab5dfafe93bb351520